msgApp/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.3
anyio==3.6.2
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
audioread==3.0.0
Babel==2.11.0
backcall==0.2.0
bleach==6.0.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
clipboard==0.0.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
comtypes==1.1.14
contourpy==1.0.7
customtkinter==5.0.5
cycler==0.11.0
Cython==0.29.33
darkdetect==0.8.0
dataclasses-json==0.5.7
DateTime==5.1
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
docopt==0.6.2
docutils==0.19
docxcompose==1.4.0
docxtpl==0.16.5
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.87.0
filelock==3.12.0
filetype==1.2.0
Flask==2.2.3
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
h11==0.14.0
h5py==3.8.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.26.0
importlib-metadata==6.0.0
ipykernel==6.21.2
ipython==8.11.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.2.0
keras==2.11.0
keyring==23.13.1
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kiwisolver==1.4.4
lazy_loader==0.3
libclang==15.0.6.1
librosa==0.10.0.post2
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
Ming==0.13.0
mongoengine==0.27.0
more-itertools==9.1.0
motor==3.1.2
MouseInfo==0.1.3
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
mwan==1.0
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
nest-asyncio==1.5.6
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.13.1
openai==0.27.8
opencv-python==4.7.0.68
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.1
parso==0.8.3
pdfkit==1.0.0
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.13
pkginfo==1.9.6
platformdirs==3.0.0
playsound==1.3.0
playwright==1.31.1
pooch==1.6.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.4
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.10.5
pydeck==0.8.0
pydub==0.25.1
pyee==9.0.4
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.14.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
PyJWT==2.6.0
PyMatting==1.1.8
pymodm==0.4.3
pymongo==3.13.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyreadline3==3.4.1
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
python-dateutil==2.8.2
python-docx==0.8.11
python-engineio==4.7.1
python-multipart==0.0.6
python-tiktok==0.1.6
pyttsx3==2.90
pytube==12.1.2
pytweening==1.0.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==305
pywin32-ctypes==0.2.0
pyzmq==25.0.0
readme-renderer==37.3
rembg==2.0.30
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.1
roman==3.3
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.9.3
scrape==0.11.3
secure-smtplib==0.1.1
semver==2.13.0
simple-websocket==0.10.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.5
SpeechRecognition==3.10.0
stack-data==0.6.2
starlette==0.21.0
streamlit==1.25.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.2.27
TikTokApi==5.2.2
tldextract==3.4.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
twilio==7.16.2
twine==4.0.2
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
yarg==0.1.9
yarl==1.8.2
zipp==3.12.1
zope.interface==6.0
